requirements.txt
opentelemetry-api
opentelemetry-sdk
tornado